feat: implementation without circular dependency
