+++ packages/contracts-sdk/src/lib/contracts-sdk.ts
@@ -55,10 +55,11 @@ import * as stakingBalancesContract from '../abis/StakingBalances.sol/StakingBal
 import { TokenInfo, derivedAddresses } from './addresses';
 import { IPubkeyRouter } from '../abis/PKPNFT.sol/PKPNFT';
 import { computeAddress } from 'ethers/lib/utils';
-import { getAuthIdByAuthMethod } from '@lit-protocol/lit-auth-client';
+import { getAuthIdByAuthMethod, stringToArrayify } from './auth-utils';
 import { Logger, LogManager } from '@lit-protocol/logger';
 import {
   calculateUTCMidnightExpiration,
+  convertRequestsPerDayToPerSecond,
   requestsToKilosecond,
 } from './utils';
 import {
@@ -1185,7 +1186,7 @@ https://developer.litprotocol.com/v3/sdk/wallets/auth-methods/#auth-method-scope
   ): Promise<MintWithAuthResponse<ContractReceipt>> => {
     const authMethodId =
       typeof params.authMethodId === 'string'
-        ? this.stringToArrayify(params.authMethodId)
+        ? stringToArrayify(params.authMethodId)
         : params.authMethodId;
 
     return this.mintWithAuth({
@@ -1225,7 +1226,7 @@ https://developer.litprotocol.com/v3/sdk/wallets/auth-methods/#auth-method-scope
   }): Promise<ethers.ContractReceipt> => {
     const _authMethodId =
       typeof authMethodId === 'string'
-        ? this.stringToArrayify(authMethodId)
+        ? stringToArrayify(authMethodId)
         : authMethodId;
 
     const _webAuthnPubkey = webAuthnPubkey ?? '0x';
@@ -2459,19 +2460,4 @@ https://developer.litprotocol.com/v3/sdk/wallets/auth-methods/#auth-method-scope
       // },
     },
   };
-
-  /**
-   * Converts a string into a byte array (arrayified value)
-   * @param str - The input string to be converted.
-   * @returns A Uint8Array representing the arrayified value of the string.
-   */
-  private stringToArrayify(str: string): Uint8Array {
-    try {
-      // Convert the string to a UTF-8 encoded byte array
-      const encoder = new TextEncoder();
-      return encoder.encode(str);
-    } catch (e) {
-      throw new Error(`Error converting string to arrayify: ${e}`);
-    }
-  }
 }
